+++ Ask.Network/solana/programs/ask-network/src/state/mod.rs
@@ -1,10 +1,14 @@
 pub use ask::*;
+pub use commercial_register::*;
+pub use commercial_register_germany::*;
 pub use global::*;
 pub use membership::*;
 pub use organization::*;
 pub use user::*;
 
 mod ask;
+mod commercial_register;
+mod commercial_register_germany;
 mod global;
 mod membership;
 mod organization;
